created separate test class for each main class -- for better testing granularity
