+++ src/test/groovy/MethodTests.groovy
@@ -1,0 +1,42 @@
+package net.vveitas.offernet
+
+import org.apache.log4j.PropertyConfigurator
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
+
+import static org.junit.Assert.*
+import static org.hamcrest.CoreMatchers.instanceOf;
+
+import org.junit.Test;
+import org.junit.BeforeClass;
+import org.junit.AfterClass;
+
+import org.apache.log4j.PropertyConfigurator
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
+
+import akka.actor.Props
+import akka.actor.ActorSystem;
+import akka.actor.ActorRef;
+
+import akka.testkit.TestActorRef
+import akka.testkit.JavaTestKit;
+
+public class MethodTests {
+	static private OfferNet on = new OfferNet().flushVertices();
+    static private Logger logger;
+    static ActorSystem system = ActorSystem.create();
+
+	@BeforeClass
+	static void initLogging() {
+	    def config = new ConfigSlurper().parse(new File('configs/log4j-properties.groovy').toURL())
+		PropertyConfigurator.configure(config.toProperties())
+		logger = LoggerFactory.getLogger('Tests.class');
+	}
+
+	@Test
+	void sendMethod() {
+		def agent1 = TestActorRef.create(system, Agent.props(on.session)).underlyingActor();
+	}
+
+}
